requirements.txt
argon2-cffi==21.1.0
asgiref==3.5.0
autopep8==1.6.0
boto3==1.18.63
botocore==1.21.65
braintree==4.13.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-environ==0.8.1
django-filter==21.1
django-heroku==0.3.1
django-js-asset==2.0.0
django-storages==1.12.2
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
Markdown==3.3.6
MarkupSafe==2.0.1
mutagen==1.45.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.0
psycopg2==2.9.2
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.1.0
pyparsing==3.0.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0